Creando fichero de configuraciónde BD y actualizando const Tabla en app.js y carga de registros iniciales
