+++ app.js
@@ -5,74 +5,8 @@ const app = express();
 const path = require('path');
 const expressLayouts = require('express-ejs-layouts');
 
-//Conexión con MongoDB
-const util = require('util');
-const mongoose = require('mongoose');
-
-mongoose.connect('mongodb://localhost/Practica9');
-const TablaSchema = mongoose.Schema({
-    "entrada_actual": String,
-    "nombre": String,
-    "descripcion": String
-    //"id": String
-});
-const Tabla = mongoose.model("Tabla",TablaSchema);
-//Ejemplos por defecto
-let ejemplo1 = new Tabla(
-{
-    entrada_actual: '"producto", "precio"\n "cam", "4,3" \n"libro de O\'Reilly", "7,2"',
-    nombre: "Ejemplo1",
-    descripcion: "Primer ejemplo para que el usuario cargue en la app"
-    //id: "Tabla1"
-});
-let p1 = ejemplo1.save(function(err)
-{
-    if(err)
-    {
-        console.log(`Hubieron errores:\n${err}`); return err; 
-    }
-    else
-    {
-        console.log(`Saved: ${ejemplo1}`);
-    }
-});
-let ejemplo2 = new Tabla(
-{
-    entrada_actual: '"producto","precio","fecha"\n "camisa","4,3","14/01"\n "libro de O\'Reilly", "7,2","13/02"',
-    nombre: "Ejemplo2",
-    descripcion: "Segundo ejemplo para que el usuario cargue en la app"
-    //id: "Tabla2"
-});
-let p2 = ejemplo2.save(function(err)
-{
-    if(err)
-    {
-        console.log(`Hubieron errores:\n${err}`); return err; 
-    }
-    else
-    {
-        console.log(`Saved: ${ejemplo2}`);
-    }
-});
-let ejemplo3 = new Tabla(
-{
-    entrada_actual: '"edad",  "sueldo",  "peso"\n  ,"6000€","90Kg"\n47,       "3000€",  "100Kg"',
-    nombre: "Ejemplo3",
-    descripcion: "Tercer ejemplo para que el usuario cargue en la app"
-    //id: "Tabla3"
-});
-let p3 = ejemplo3.save(function(err)
-{
-    if(err)
-    {
-        console.log(`Hubieron errores:\n${err}`); return err; 
-    }
-    else
-    {
-        console.log(`Saved: ${ejemplo3}`);
-    }
-});
-
+//Conexión con Estructura de MongoDB
+const Tabla = require('./models/estructura_bd.js');
 
 app.set('port', (process.env.PORT || 5000));
 
@@ -100,6 +34,10 @@ app.get('/guardar_tabla',(request, response) => {
     //Comprobamos el número de registros en la base de datos
     Tabla.find({},function(err, data) 
     {
+        if(err)
+        {
+            console.error("Se ha producido un error->"+err);
+        }
         console.log("Data.length --> "+data.length);
         if(data.length > 3)
         {
@@ -108,7 +46,7 @@ app.get('/guardar_tabla',(request, response) => {
         }
         let nueva_tabla = new Tabla(
         {
-            entrada_actual: request.query.input,
+            entrada_tabla: request.query.input,
             nombre: request.query.nombre,
             descripcion: request.query.descripcion
         });
@@ -125,11 +63,12 @@ app.get('/guardar_tabla',(request, response) => {
                 console.log(`Saved: ${nueva_tabla}`);
             }
             
-            //Nos aseguramos de que todos los registros se han salvado
-            Promise.all([n_t]).then( (value) => {
-                console.log(util.inspect(value, {depth: null}));
-            });        
         });
+        
+        //Nos aseguramos de que todos los registros se han salvado
+        Promise.all([n_t]).then( (value) => {
+            console.log(util.inspect(value, {depth: null}));
+        }); 
     });
 });
