--- public/Ejemplo1.txt
@@ -1,3 +1,0 @@
-"producto",           "precio"
-"camisa",             "4,3"
-"libro de O\"Reilly", "7,2"
