--- public/Ejemplo2.txt
@@ -1,3 +1,0 @@
-"producto",           "precio"  "fecha"
-"camisa",             "4,3",    "14/01"
-"libro de O\"Reilly", "7,2"     "13/02"
