--- public/Ejemplo3.txt
@@ -1,3 +1,0 @@
-"edad",  "sueldo",  "peso"
-,         "6000€",  "90Kg"
-47,       "3000€",  "100Kg"
